Restyled View List and added edit functionality directly, deleted Edit List
