+++ client/src/index.js
@@ -6,7 +6,6 @@ import reportWebVitals from './reportWebVitals';
 import { App } from './components/App';
 import { CreateList } from './components/CreateList';
 import { ViewList } from './components/ViewList';
-import { EditList } from './components/EditList';
 import { Find } from './components/Find';
 import { NewPlace } from './components/NewPlace';
 import { Menu } from './components/Menu';
@@ -18,7 +17,6 @@ ReactDOM.render(
       <Switch>
         <Route path="/" exact component={App} />
         <Route path="/lists/new" exact component={CreateList} />
-        <Route path="/lists/edit" exact component={EditList} />
         <Route path="/lists/:id" exact component={ViewList} />
         <Route path="/find" exact component={Find} />
         <Route path="/places/new" exact component={NewPlace} />
