+++ client/src/styles/Lists.css
@@ -1,7 +1,7 @@
-.card__min_width {
+.card_min_width {
   width: 300px;
 }
 
-.card__background_color {
-  background: rgba(33, 150, 243, 0.6);
+.card_background_color {
+  background-color: rgba(33, 150, 243, 0.6);
 }
