+++ client/src/styles/ViewList.css
@@ -1,0 +1,7 @@
+.datatable_max_width {
+  max-width: 600px;
+}
+
+.button_float_right {
+  float: right
+}
